requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.30.1
anthropic>=0.40.0
python-dotenv==1.0.1
jinja2==3.1.4
schedule==1.2.1
pyyaml==6.0.1

